recommendation_engine/requirements.txt
pandas>=2.1.0
numpy>=1.25.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
